Move check cfg tests to their own module
